Impl constructors and findKmConstructor
